refactor(diff): move to separate internal package
